item add delete view

+++ src/main/java/com/example/demo/Entity/Item.java
@@ -4,7 +4,6 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-
 import lombok.Data;
 
 @Entity
@@ -12,13 +11,10 @@ import lombok.Data;
 public class Item {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-private Integer itemid;
-private	float mrp ;
-private	float discount; 
-private  float price ;
+private Long itemid;
+private float mrp;
+private float discount;
+private float price;
 private String created_at;
-private String update_at; 
-	
-	
-	
+private String update_at;	
 }
