+++ src/main/java/com/example/demo/Model/ItemDTO.java
@@ -4,10 +4,10 @@ import lombok.Data;
 
 @Data
 public class ItemDTO {
-	private Integer itemid;
-	private	float mrp ;
-	private	float discount; 
-	private  float price ;
+	private Long itemid;
+	private float mrp;
+	private float discount;
+	private float price;
 	private String created_at;
-	private String update_at; 
+	private String update_at;	 
 }
